refactor github crawling
