+++ backend/src/main/java/com/beyond/specguard/common/config/GitHubRestTemplateConfig.java
@@ -1,19 +1,24 @@
-package com.github.github.config;
+package com.beyond.specguard.common.config;
 
-import org.springframework.beans.factory.annotation.Value;
+import com.beyond.specguard.common.properties.AppProperties;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
 
+@Slf4j
 @Configuration
-public class RestTemplateConfig {
+@RequiredArgsConstructor
+public class GitHubRestTemplateConfig {
 
-    @Value("{GITHUB_TOKEN}")
-    private String githubToken;
+    private final AppProperties appProperties;
 
     @Bean
     public RestTemplate githubRestTemplate(RestTemplateBuilder builder) {
+        String githubToken = appProperties.getGithub().getToken();
+
         return builder
                 .defaultHeader("Authorization", "token " + githubToken)
                 .defaultHeader("Accept", "application/vnd.github.v3+json")
