+++ backend/src/main/java/com/beyond/specguard/common/properties/AppProperties.java
@@ -13,6 +13,7 @@ public class AppProperties {
 
     private Jwt jwt = new Jwt();
     private Redis redis = new Redis();
+    private Github github = new Github();
 
     @Getter @Setter
     public static class Jwt {
@@ -31,5 +32,10 @@ public class AppProperties {
             private String blacklist;
             private String session;
         }
+    }
+
+    @Getter @Setter
+    public static class Github {
+        private String token;
     }
 }
